+++ applications/desktop/src/common/config/index.ts
@@ -1,0 +1,4 @@
+export { ALL_CONFIG_OTIONS } from "./config-options";
+export * from "./use-cases";
+export * from "./paths";
+export * from "./schema";
