+++ applications/desktop/src/preferences/index.tsx
@@ -3,7 +3,6 @@
  */
 
 import { DarkTheme, LightTheme } from "@nteract/presentational-components";
-import { AppState } from "@nteract/types";
 import * as React from "react";
 import ReactDOM from "react-dom";
 import { connect, Provider } from "react-redux";
