+++ applications/desktop/src/preferences/items/index.tsx
@@ -1,17 +1,11 @@
 import * as React from "react";
 
-import { ConfigHeading, Heading, isHeading } from "./heading";
-import { BooleanOption, ConfigOptionBoolean, isBoolean } from "./option-boolean";
-import { ConfigOptionEnum, EnumOption, isEnum } from "./option-enum";
-import { ConfigOptionKernels, isKernels, KernelsOption } from "./option-kernels";
+import { ConfigItem } from "../../common/config/schema";
+import { Heading, isHeading } from "./heading";
+import { BooleanOption, isBoolean } from "./option-boolean";
+import { EnumOption, isEnum } from "./option-enum";
+import { isKernels, KernelsOption } from "./option-kernels";
 
-export type ConfigOption =
-  | ConfigOptionEnum
-  | ConfigOptionBoolean
-  | ConfigOptionKernels
-  ;
-
-export type ConfigItem = ConfigHeading | ConfigOption;
 
 export const Item = (props: ConfigItem) =>
   <React.Fragment>
